+++ src/dialog/cueinterpreterdialog.cpp
@@ -7,6 +7,11 @@
 #include <QSettings>
 #include <QString>
 
+/**
+ *	TODO: Add ability to edit Cues
+ *  - Allow users to load and save Cuefiles from the dialog
+ *  - Users can delete Cues (e.g. remove a line)
+ */
 namespace PixelMaestroStudio {
 	CueInterpreterDialog::CueInterpreterDialog(QWidget *parent, uint8_t* cuefile, uint32_t size) : QDialog(parent), ui(new Ui::CueInterpreterDialog), model_(cuefile, size) {
 		ui->setupUi(this);