+++ src/dialog/paletteeditdialog.cpp
@@ -9,6 +9,7 @@
 #include "ui_paletteeditdialog.h"
 #include "widget/palettecontrolwidget.h"
 
+// TODO: Allow editing color ranges as opposed to having to edit an entire Palette all at once
 namespace PixelMaestroStudio {
 	/**
 	 * Constructor.